requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.2
cycler==0.10.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.2
mysql-connector-python==8.0.25
numpy==1.21.1
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
pyTelegramBotAPI==3.8.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
